chore: move timer logic to hook
